Move MUltiThreadedFunSuite into a spinal package
