+++ src/test/scala/naxriscv/NaxRiscvRegression.scala
@@ -11,6 +11,7 @@ import org.apache.commons.io.FileUtils
 import org.scalatest.Tag
 import org.scalatest.funsuite.AnyFunSuite
 import spinal.core.SpinalConfig
+import spinal.lib.misc.test.MultithreadedFunSuite
 
 import java.io.{File, OutputStream}
 import java.util.concurrent.ForkJoinPool
@@ -20,50 +21,7 @@ import scala.concurrent.duration.Duration
 import scala.sys.process.{Process, ProcessLogger}
 import scala.util.Random
 
-class MultithreadedFunSuite(threadCount : Int) extends AnyFunSuite {
-  val finalThreadCount = if(threadCount > 0) threadCount else {
-    new oshi.SystemInfo().getHardware.getProcessor.getLogicalProcessorCount
-  }
-  implicit val ec = ExecutionContext.fromExecutorService(
-    new ForkJoinPool(finalThreadCount, ForkJoinPool.defaultForkJoinWorkerThreadFactory, null, true)
-  )
-  class Job(body : => Unit){
-    val originalOutput = Console.out
-    val buffer = mutable.Queue[Char]()
-    var bufferEnabled = true
-    def redirector() = new OutputStream{
-      override def write(i: Int): Unit = synchronized {
-        if(bufferEnabled) buffer += i.toChar
-        else originalOutput.print(i.toChar)
-      }
-    }
-    val future = Future{
-      Console.withOut(redirector()){
-        Console.withErr(redirector())(body)
-      }
-    }
-
-    def join(): Unit = {
-      Thread.sleep(50)
-      synchronized{
-        bufferEnabled = false
-        buffer.foreach(originalOutput.print)
-      }
-      Await.result(future, Duration.Inf)
-    }
-  }
-
-  def testMp(testName: String, testTags: Tag*)(testFun: => Unit) {
-    val job = new Job(testFun)
-    super.test(testName, testTags :_*)(job.join())
-  }
-  protected def testSingleThread(testName: String, testTags: Tag*)(testFun: => Unit) {
-    super.test(testName, testTags :_*)(testFun)
-  }
-}
-
 object HeavyLock
-
 class NaxRiscvRegression extends MultithreadedFunSuite(sys.env.getOrElse("NAXRISCV_REGRESSION_THREAD_COUNT", "4").toInt){
 
   var seed = sys.env.getOrElse("NAXRISCV_SEED", Random.nextInt(100000000).toString).toInt
